Created own class for Leaderboard
